+++ CHANGELOG.md
@@ -11,9 +11,12 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
-- Add possibility of switcing between random and sorted data
+- Add possibility of switching between random and sorted data
+- Fixed that /random should not show not accepted tiles
+- Fix colored overlay
 - Added monolog and mime packages
-- Fix tiles image path
+- Fixed tiles image path
+- Fixed defaults when tile.extra or screen.variant are not parsable.
 
 ### Added
 - Ctabox in frontend